Add yaml_file_stored_as var to project.yaml files and re-org dirs
In all the different old tests to build up the nectar VM to run bdpg,
it's hard to figure out which version of project.yaml and model.R files
were used.  So, I'm now saving an appropriately named copy of each
and adding variables and output in the model.R and project.yaml files
to echo that information at runtime.  This should make it a little bit
easier when looking at a run to see what went into it.