Replace model/yaml pairs directory with just modelRFiles directory
There are really only 2 model.R files, one that calls the bdpg code
and one that doesn't.  The one that doesn't is only used for the
very first test.  All other runs use the second one, so it's pointless
to have a directory of all model.R and project.yaml pairs since only
one of them uses the first model.R and there is already a directory
storing all of the project.yaml versions in order of use.