Update test 17y to do 13 more runs since at least 13 failed
These failures seem to have been due to a loss of connection to the gurobi
license server from one of the nectar VMs during a particular small time
window.  They all finished in 2 or 3 minutes instead of 10 or 20 hours and
all of them choked on the gurobi call.  Should be sufficient to just do
13 more runs since that machine and all the other have had no problems since
that one time window.