Fix missing designation of gen_basic_variants_err_amt
This is a repair after the fact since all 100 runs of a hard problem at
error level 0.05 have already been done.  It's just that in getting ready
for the 0.02 error level, I realized that somewhere along the way, the
line setting the error level was removed.  It didn't matter, because
it defaults to 0.05 in the code and that was the intended error level
anyway.  However, for consistency and clarity, it's better to have this
line in the code since all levels other than 0.05 need to set it explicitly.