Undo previous fix of bipartite metric choice for the moment
Fixing the bipartite metric choice before I have finished all of the 100 hard
run error levels on nectar will mean that the output files will have different
columns.  So, I'm going to put the mistake back in and finish out all of the
hard run error levels.  After that, I'm going to have to redo all of these
runs for both easy and hard to get the bipartite metric values...