Temp hack to yaml to correct desired_Xu_spp_frac_of_all_spp
bdpg had a bug in using the desired_Xu_spp_frac_of_all_spp that caused the
proportions for this and the fraction that drives the num_spp_to_generate
for the wrapped species to be reversed.

In all previous runs, the desired value had been set to 50%, so the bug had
no effect since both proportions would be the same.

I have fixed the bug at bdpg commit de06807 but using that code on nectar
will require reinstalling the bdpg package by hand on 20 different nectar
machines right now.  So, for the moment I'm just going to reverse the
proportion specified in the yaml file while I continue to do some more
small tests on nectar.  As soon as the bdpg fix is installed on nectar,
then the value for desired_Xu_spp_frac_of_all_spp in here needs to be
replaced with 1 minus it's current hack value.