Set yaml vars for first attempt at 500 runs of all but gurobi time limited
Looks things are ready for trying to generate the first 500 experiments
on nectar for the random Xu problems with 6 variants, all reserve selectors,
and all but slowest graph metrics.  The only other limitation is that
gurobi is time limited to however long marxan takes on each problem so
that gurobi doesn't slow everything else down.  We can go back and rerun
any gurobi problems later if necessary.  In the last test run, only 13%
of the gurobi runs failed to reach their optimum in the marxan time limit.