Change yaml to rerun 1 previously failed run using same seed.
Setting up to rerun 8 runs that ran out of memory so that they can rerun on
larger nectar machines.  However, at the moment, this has to be done 1 at
a time because there's no code yet to run a whole a whole group of
runs using their original random seed.  Want to get them working while I
go back and modify the code, so I have to commit a separate yaml file for
each of the 8 reruns and schedule each of them separately.