First try at using yaml array of random seeds on nectar
A few runs have failed due to lack of memory for large problems, so
I've added code to bdpg to allow the yaml file to contain an array of
random seeds taken from the failed runs and then rerun those runs using
the seed array.  They should produce exactly the same run as before, but
this time they will be running on larger machines so they should run to
completion.  This yaml file is just trying 7 of the first 8 failures
(i.e., those from Batch 1).  The very first failure is not included because
it has already been rerun on its own in the previous incarnation of
this yaml file before I had written the code for handling arrays of seeds.