Set up to do 1 more run for second exp batch that only had 499 runs.
I think that I had accidentally killed all the small nectar machines while
there was still 1 run in progress, so only 499 were done instead of 500.
This run is not one of the failed ones that were handled in the previous
commit, but will run it on a large machine as if it was a failed run and
that should guarantee that it does not fail.