Replaced yaml array with edited copy from old version
Still trying to fix crashing when using the yaml array of random seeds.
This time I'm taking the array from an old project.yaml file, then editing
the random seeds in that array by pasting in the new seeds after having
them in TextEdit instead of TextWrangler.  This time, git shows them as
being different, so maybe this will help...