Set up for batch 5 exps going back to fixed Xu spp frac for wraps
Batch 4 used a random value for desired_Xu_spp_frac_of_all_spp between 0.25
and 0.5, but that led to almost no larger problems being generated.
Where earlier batches had 10 or 15 problems that had to be rerun on large
machines, batch 4 only had 1.  Since we want to explore behavior on larger,
harder problems too, I'm going back to the fixed value of 0.25 for these
experiments.  This level gives Some bigger problems, but not so many that
they lock up all the machines all the time.  It's about as many as I can
do in reasonable time given the number of machines in the current nectar
allocation, since the larger problems are taking going at about 2 per day
 on the large marchines while all other problems can be done at the rate of
nearly 500 problems in 2 days.