Set up for 2 Batch 3 reruns on m3.large machines
The second of these reruns is itself a rerun that ran out of memory on
a large machine already and is the reason that I'm having to again
step up the size of the machine used.
The first of the two has only failed on the small machine so it probably
doesn't need the bigger machine for its rerun, but I'll let it run on the
same m3.large machine setup since it's from the same batch 3 and I
have some spare capacity at the moment and it's easier to include it in
the same yaml file rather than making 2 separate ones.